Move browse related code to its own controller
